sparkle-races-angular/src/app/sign-in/sign-in.component.ts
import { Component, OnInit } from '@angular/core';
import { AuthService } from '../auth.service';
import { NgForm } from '@angular/forms';
import { Router } from '@angular/router';

@Component({
  selector: 'app-sign-in',
  templateUrl: './sign-in.component.html',
  styleUrls: ['./sign-in.component.scss']
})
export class SignInComponent implements OnInit {

  constructor(private authService: AuthService, private router: Router) { }

  ngOnInit(): void {
  }

onSignIn(form: NgForm){

  this.authService.signIn(form.value).subscribe(() => {

    this.router.navigate(['/'])
  }
  )
}

}
